2016.04.22_lesson18/django_orm_example/requirements.txt
astroid==1.4.4
colorama==0.3.6
Django==1.9.5
django-debug-toolbar==1.4
django-querycount==0.4.1
django-split-settings==0.1.3
docopt==0.6.2
gitdb==0.6.4
GitPython==1.0.2
joe==0.0.7
lazy-object-proxy==1.2.1
mercurial==3.6.3
pbr==1.8.1
pep8==1.7.0
powerline-status==2.3
pylint==1.5.4
pytz==2016.3
six==1.10.0
smmap==0.9.0
sqlparse==0.1.19
stevedore==1.10.0
virtualenv==14.0.5
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.26.0
wrapt==1.10.6
